+++ NEWS.md
@@ -39,6 +39,9 @@
 ## MAJOR CHANGES
 
 - Upgraded the DataTables library to 1.10.11; there have been many changes from 1.10.7 to this version: http://datatables.net/blog/2015-08-13
+    - The `extensions` argument of `datatable()` should be a character vector now; previously it can be a list of initialization options for extensions due to the inconsistent ways of initializing DataTables extensions; now the initialization options for all extensions can be set in the `options` argument.
+    - The `copySWF()` function has been removed, since the `TableTools` extension has been removed.
+    - The `ColVis` extension was removed and replaced but the `colvis` button in the `Buttons` extension.
 
 - In the previous version, row names were used as row indices in the server-side processing mode, but numeric row indices were used in the client-side mode. Now we always use numeric row indices in both modes for the sake of consistency. These input values in Shiny will always be integers: `input$tableId_rows_current`, `input$tableId_rows_all`, and `input$tableId_rows_selected`.